pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jobflow"
version = "0.1.0"
description = "CLI tool to tailor resumes for job postings"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "pyyaml>=6.0",
]

[project.scripts]
jobflow = "jobflow.cli:app"

[tool.setuptools.packages.find]
include = ["jobflow*"]
